+++ ChangeLog.md
@@ -6,6 +6,7 @@
 - Implement method `myDataset::load()` to load mnist dataset.
 - Add method into the class `myDataset` to get the fields.
 - Add the class `myDatasetIterator` to travel a whole dataset.
+- Add the class `myNetwork` to predict.
 
 ### Changed
 - Update the class `myDatasetIterator` for batch training.